migrate database to include new BIO data removed old scripts (importAll) import new BIO data fixed voteDate override to be more accurate
