--- src/express/webhooks/debug/importAll.js
@@ -1,3 +1,0 @@
-import importAll from '../../../scripts/importAll';
-
-module.exports = importAll;
