+++ src/graphql/queries/getProcedureUpdates.js
@@ -36,6 +36,7 @@ export default gql`
         }
         namedVote
         voteDate
+        voteEnd
         customData {
           voteResults {
             yes
@@ -53,6 +54,19 @@ export default gql`
                 no
                 abstination
                 notVoted
+              }
+            }
+          }
+        }
+        sessions {
+          thisYear
+          thisWeek
+          session {
+            session
+            top {
+              heading
+              topic{
+                isVote
               }
             }
           }
