+++ src/models/Procedure.js
@@ -1,14 +1,4 @@
 import mongoose from 'mongoose';
-import ProcedureSchema from './../migrations/3-schemas/Procedure';
+import ProcedureSchema from './../migrations/7-schemas/Procedure';
 
 export default mongoose.model('Procedure', ProcedureSchema);
-
-/*
-mongoose.model('Procedure').createIndexes(err => {
-  if (!err) {
-    Log.info('SearchIndexs for Procedures created');
-  } else {
-    Log.error(JSON.stringify({ err }));
-  }
-});
-*/
